requirements.txt
requests
bs4
## 
selenium
browsermob-proxy
psutils